requirement.txt
pypdf
langchain_community
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
chainlit
huggingface_face
langchain
CTransformers
chainlit